Initial attempt to isolate Client/Server code from types and classes
